requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
langchain-groq
pypdf
